modified test files

+++ application/__init__.py
@@ -11,18 +11,25 @@ app = Flask(__name__)
 from application.build_database import db_load_example_data
 
 
+# create application instance with application factoy pattern
 def create_app(configs_string):
     app.config.from_object(configs_string)
-
     db.init_app(app)
+
+    # configure logging of application events
     setup_app_logging(app)
 
+    # importing and registering app blueprints
     register_blueprints()
+
     # deleting pre-existing database for the tests in case test configuration is passed as a parameter
-    with app.app_context():
-        if os.path.exists("../tests/test.db"):
-            os.remove("../tests/test.db")
-        db.create_all()
+    if configs_string == 'settings.TestConfig':
+        with app.app_context():
+            if os.path.exists("../tests/test.db"):
+                os.remove("../tests/test.db")
+
+    # creating and populating a database with sample data
+    db.create_all()
     db_load_example_data(app, db)
 
     return app
@@ -43,4 +50,4 @@ def setup_app_logging(app):
     app.logger.addHandler(log_handler_file)
     if not app.debug:
         logging.basicConfig(filename='appevents.log', level=logging.WARNING,
-                        format='%(levelname)s:%(message)s')
+                            format='%(levelname)s:%(message)s')
